// src/db/enums.ts
// AccountType enum
export enum AccountType {
  CUSTOMER = 'CUSTOMER',
  PROVIDER = 'PROVIDER'
}

// PaymentRule enum
export enum PaymentRule {
  DEFAULT = 'DEFAULT',
  FIRST_MONDAY = 'FIRST-MONDAY',
  FIRST_MON_NORFND = 'FIRST-MON-NORFND',
  FIRST_MON_CANCEL = 'FIRST-MON-CANCEL',
  FNIGHT_CHECKIN = 'FNIGHT-CHECKIN',
  FNIGHT_CHKIN_5 = 'FNIGHT-CHKIN+5',
  FNIGHT_CHKIN_10 = 'FNIGHT-CHKIN+10',
  FNIGHT_CHKIN_15 = 'FNIGHT-CHKIN+15',
  FNIGHT_CHKIN_5_NORFND = 'FNIGHT-CHKIN+5-NORFND',
  FNIGHT_CHKIN_10_NORFND = 'FNIGHT-CHKIN+10-NORFND',
  FNIGHT_CHKIN_15_NORFND = 'FNIGHT-CHKIN+15-NORFND',
  FNIGHT_CHKIN_0_NORFND = 'FNIGHT-CHKIN+0-NORFND',
  FNIGHT_AUTOCANCEL = 'FNIGHT-AUTOCANCEL',
  WEEK_CHECKIN_7 = 'WEEK-CHECKIN+7',
  WEEK_CHECKOUT = 'WEEK-CHECKOUT',
  TENS_CHECKIN_7 = 'TENS-CHECKIN+7',
  NEXT_MONTH = 'NEXT-MONTH',
  CANPOLICIES_7 = 'CANPOLICIES-7',
  CANPOLICIES_4 = 'CANPOLICIES-4',
  CANPOLICIES_3 = 'CANPOLICIES-3',
  CANPOLICIES_2 = 'CANPOLICIES-2',
  CANPOLICIES_1 = 'CANPOLICIES-1',
  AUTOCANCEL_7 = 'AUTOCANCEL-7',
  AUTOCANCEL_6 = 'AUTOCANCEL-6',
  AUTOCANCEL_5 = 'AUTOCANCEL-5',
  AUTOCANCEL_4 = 'AUTOCANCEL-4',
  AUTOCANCEL_3 = 'AUTOCANCEL-3',
  AUTOCANCEL_2 = 'AUTOCANCEL-2',
  AUTOCANCEL_1 = 'AUTOCANCEL-1',
  CHECKIN_PLUS_1 = 'CHECKIN+1',
  CHECKIN_PLUS_2 = 'CHECKIN+2',
  CHECKIN_PLUS_29 = 'CHECKIN+29',
  CHECKIN_MINUS_1 = 'CHECKIN-1',
  CHECKIN_MINUS_4 = 'CHECKIN-4',
  CHECKIN_MINUS_7 = 'CHECKIN-7'
}